Filtering and ordering CaseViewset. Fixes some things around the creating case
