+++ app/apps/cases/serializers/__init__.py
@@ -1,0 +1,3 @@
+from .bwv import *  # noqa
+from .case import *  # noqa
+from .main import *  # noqa
