+++ e2e-tests/api/tasks/close_case.py
@@ -34,7 +34,7 @@ class test_afsluiten_zaak(AbstractUserTask, task_afsluiten_zaak):
 
     def get_post_data(self, case, task):
         return super().get_post_data(case, task) | {
-            "theme": case.data["theme"],
+            "theme_id": case.data["theme"],
             "case_user_task_id": task["case_user_task_id"],
         }
